Set up Continuous Integration
For pull requests, this tries to build the container (which
also runs unit tests, due to how the Dockerfile is set up).
GitHub will display the build status on the review page for
the pull request with an icon that is green if the build
succeeded, or red if it failed.

For changes to the main branch, this builds the container
and pushes it to the GitHub container registry. Once this
change is part of the main repo and GitHub had a chance to
run the corresponding action for the first time, users will
be able to see the most current container from:

https://github.com/bdon/OSMExpress/pkgs/container/osmexpress

Users can then use one of the following commands to fetch it,
depending on whether they prefer docker (which needs installation
of proprietary software) or podman (which is an purely open-source
implementation):

$ docker pull ghcr.io/bdon/osmexpress:latest
$ podman pull ghcr.io/bdon/osmexpress:latest