Set up Continuous Integration
For pull requests, we try to build the container. This
also runs unit tests due to how the Dockerfile is set up.
GitHub will display the build status on its review page for
the pull request. If the build (and tests that are run as
part of the build) succeeded, the icon is green. If it failed,
the icon is red.

For changes to the main branch, we build the container
and push it to the GitHub container registry. Once this
change is part of the main repo and GitHub had a chance to
run the corresponding workflow for the first time, users
can use one of the following commands to fetch the latest
container, depending on whether they prefer docker (which
needs installation of proprietary software) or podman
(which is open source):

```sh
$ docker pull ghcr.io/bdon/osmexpress:latest
$ podman pull ghcr.io/bdon/osmexpress:latest
```

The full list of `osmexpress` containers will be available from
[this page](https://github.com/bdon/OSMExpress/pkgs/container/osmexpress).